requirements.txt
flask
python-dotenv
psycopg2-binary
python-dotenv
flask_restplus
flask_jwt_extended
-e .